Move IconName into metabase-types
