+++ frontend/src/metabase-types/api/card.ts
@@ -2,8 +2,8 @@ import type {
   EmbeddingParameters,
   EmbeddingType,
 } from "metabase/embedding/types";
-import type { IconName } from "metabase/ui";
 import type { CurrencyStyle } from "metabase/utils/formatting";
+import type { IconName } from "metabase-types/api";
 import type { EntityToken, EntityUuid } from "metabase-types/api/entity";
 
 import type { ClickBehavior } from "./click-behavior";
